+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_api_impls.dart
@@ -11,7 +11,8 @@ import 'android_webview.dart';
 import 'android_webview.g.dart';
 import 'instance_manager.dart';
 
-export 'android_webview.g.dart' show FileChooserMode;
+export 'android_webview.g.dart'
+    show ConsoleMessage, ConsoleMessageLevel, FileChooserMode;
 
 /// Converts [WebResourceRequestData] to [WebResourceRequest]
 WebResourceRequest _toWebResourceRequest(WebResourceRequestData data) {
@@ -552,6 +553,11 @@ class WebSettingsHostApiImpl extends WebSettingsHostApi {
       enabled,
     );
   }
+
+  /// Helper method to convert instances ids to objects.
+  Future<String> getUserAgentStringFromInstance(WebSettings instance) {
+    return getUserAgentString(instanceManager.getIdentifier(instance)!);
+  }
 }
 
 /// Host api implementation for [JavaScriptChannel].
@@ -894,6 +900,17 @@ class WebChromeClientHostApiImpl extends WebChromeClientHostApi {
   }
 
   /// Helper method to convert instances ids to objects.
+  Future<void> setSynchronousReturnValueForOnConsoleMessageFromInstance(
+    WebChromeClient instance,
+    bool value,
+  ) {
+    return setSynchronousReturnValueForOnConsoleMessage(
+      instanceManager.getIdentifier(instance)!,
+      value,
+    );
+  }
+
+  /// Helper method to convert instances ids to objects.
   Future<void> setSynchronousReturnValueForOnJsAlertFromInstance(
     WebChromeClient instance,
     bool value,
@@ -1083,6 +1100,13 @@ class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
         instance,
       );
     }
+  }
+
+  @override
+  void onConsoleMessage(int instanceId, ConsoleMessage message) {
+    final WebChromeClient instance =
+        instanceManager.getInstanceWithWeakReference(instanceId)!;
+    instance.onConsoleMessage?.call(instance, message);
   }
 }