+++ packages/webview_flutter/webview_flutter_android/test/android_navigation_delegate_test.dart
@@ -523,6 +523,7 @@ class CapturingWebChromeClient extends android_webview.WebChromeClient {
     super.onShowCustomView,
     super.onHideCustomView,
     super.onPermissionRequest,
+    super.onConsoleMessage,
     super.binaryMessenger,
     super.instanceManager,
   }) : super.detached() {