+++ packages/webview_flutter/webview_flutter_android/test/android_webview_controller_test.dart
@@ -76,6 +76,9 @@ void main() {
               android_webview.CustomViewCallback callback)?
           onShowCustomView,
       void Function(android_webview.WebChromeClient instance)? onHideCustomView,
+      void Function(android_webview.WebChromeClient instance,
+              android_webview.ConsoleMessage message)?
+          onConsoleMessage,
     })? createWebChromeClient,
     android_webview.WebView? mockWebView,
     android_webview.WebViewClient? mockWebViewClient,
@@ -119,6 +122,9 @@ void main() {
                         onShowCustomView,
                     void Function(android_webview.WebChromeClient instance)?
                         onHideCustomView,
+                    void Function(android_webview.WebChromeClient instance,
+                            android_webview.ConsoleMessage message)?
+                        onConsoleMessage,
                   }) =>
                       MockWebChromeClient(),
               createAndroidWebView: () => nonNullMockWebView,
@@ -618,6 +624,7 @@ void main() {
           dynamic onPermissionRequest,
           dynamic onShowCustomView,
           dynamic onHideCustomView,
+          dynamic onConsoleMessage,
         }) {
           onShowFileChooserCallback = onShowFileChooser!;
           return mockWebChromeClient;
@@ -691,6 +698,7 @@ void main() {
           dynamic onPermissionRequest,
           dynamic onShowCustomView,
           dynamic onHideCustomView,
+          dynamic onConsoleMessage,
         }) {
           onGeoPermissionHandle = onGeolocationPermissionsShowPrompt!;
           onGeoPermissionHidePromptHandle = onGeolocationPermissionsHidePrompt!;
@@ -768,6 +776,7 @@ void main() {
               onShowCustomView,
           void Function(android_webview.WebChromeClient instance)?
               onHideCustomView,
+          dynamic onConsoleMessage,
         }) {
           onShowCustomViewHandle = onShowCustomView!;
           onHideCustomViewHandle = onHideCustomView!;
@@ -823,6 +832,7 @@ void main() {
           )? onPermissionRequest,
           dynamic onShowCustomView,
           dynamic onHideCustomView,
+          dynamic onConsoleMessage,
         }) {
           onPermissionRequestCallback = onPermissionRequest!;
           return mockWebChromeClient;
@@ -880,6 +890,7 @@ void main() {
           )? onPermissionRequest,
           dynamic onShowCustomView,
           dynamic onHideCustomView,
+          dynamic onConsoleMessage,
         }) {
           onPermissionRequestCallback = onPermissionRequest!;
           return mockWebChromeClient;
@@ -903,6 +914,107 @@ void main() {
       );
 
       expect(callbackCalled, isFalse);
+    });
+
+    test('setOnConsoleLogCallback', () async {
+      late final void Function(
+        android_webview.WebChromeClient instance,
+        android_webview.ConsoleMessage message,
+      ) onConsoleMessageCallback;
+
+      final MockWebChromeClient mockWebChromeClient = MockWebChromeClient();
+      final AndroidWebViewController controller = createControllerWithMocks(
+        createWebChromeClient: ({
+          dynamic onProgressChanged,
+          dynamic onShowFileChooser,
+          dynamic onJsAlert,
+          dynamic onJsConfirm,
+          dynamic onJsPrompt,
+          dynamic onGeolocationPermissionsShowPrompt,
+          dynamic onGeolocationPermissionsHidePrompt,
+          dynamic onPermissionRequest,
+          dynamic onShowCustomView,
+          dynamic onHideCustomView,
+          void Function(
+            android_webview.WebChromeClient,
+            android_webview.ConsoleMessage,
+          )? onConsoleMessage,
+        }) {
+          onConsoleMessageCallback = onConsoleMessage!;
+          return mockWebChromeClient;
+        },
+      );
+
+      final Map<String, JavaScriptLogLevel> logs =
+          <String, JavaScriptLogLevel>{};
+      await controller.setOnConsoleMessage(
+        (JavaScriptConsoleMessage message) async {
+          logs[message.message] = message.level;
+        },
+      );
+
+      onConsoleMessageCallback(
+        mockWebChromeClient,
+        ConsoleMessage(
+          lineNumber: 42,
+          message: 'Debug message',
+          level: ConsoleMessageLevel.debug,
+          sourceId: 'source',
+        ),
+      );
+      onConsoleMessageCallback(
+        mockWebChromeClient,
+        ConsoleMessage(
+          lineNumber: 42,
+          message: 'Error message',
+          level: ConsoleMessageLevel.error,
+          sourceId: 'source',
+        ),
+      );
+      onConsoleMessageCallback(
+        mockWebChromeClient,
+        ConsoleMessage(
+          lineNumber: 42,
+          message: 'Log message',
+          level: ConsoleMessageLevel.log,
+          sourceId: 'source',
+        ),
+      );
+      onConsoleMessageCallback(
+        mockWebChromeClient,
+        ConsoleMessage(
+          lineNumber: 42,
+          message: 'Tip message',
+          level: ConsoleMessageLevel.tip,
+          sourceId: 'source',
+        ),
+      );
+      onConsoleMessageCallback(
+        mockWebChromeClient,
+        ConsoleMessage(
+          lineNumber: 42,
+          message: 'Warning message',
+          level: ConsoleMessageLevel.warning,
+          sourceId: 'source',
+        ),
+      );
+      onConsoleMessageCallback(
+        mockWebChromeClient,
+        ConsoleMessage(
+          lineNumber: 42,
+          message: 'Unknown message',
+          level: ConsoleMessageLevel.unknown,
+          sourceId: 'source',
+        ),
+      );
+
+      expect(logs.length, 6);
+      expect(logs['Debug message'], JavaScriptLogLevel.debug);
+      expect(logs['Error message'], JavaScriptLogLevel.error);
+      expect(logs['Log message'], JavaScriptLogLevel.log);
+      expect(logs['Tip message'], JavaScriptLogLevel.debug);
+      expect(logs['Warning message'], JavaScriptLogLevel.warning);
+      expect(logs['Unknown message'], JavaScriptLogLevel.log);
     });
 
     test('runJavaScript', () async {
@@ -1174,6 +1286,20 @@ void main() {
       verify(mockWebView.settings).called(1);
       verify(mockSettings.setUserAgentString('Test Framework')).called(1);
     });
+
+    test('getUserAgent', () async {
+      final MockWebSettings mockSettings = MockWebSettings();
+      final AndroidWebViewController controller = createControllerWithMocks(
+        mockSettings: mockSettings,
+      );
+
+      const String userAgent = 'str';
+
+      when(mockSettings.getUserAgentString())
+          .thenAnswer((_) => Future<String>.value(userAgent));
+
+      expect(await controller.getUserAgent(), userAgent);
+    });
   });
 
   test('setMediaPlaybackRequiresUserGesture', () async {
@@ -1361,6 +1487,7 @@ void main() {
                   android_webview.CustomViewCallback callback)?
               onShowCustomView,
           dynamic onHideCustomView,
+          dynamic onConsoleMessage,
         }) {
           onShowCustomViewCallback = onShowCustomView;
           return mockWebChromeClient;