+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/FWFGeneratedWebKitApis.m
@@ -2466,19 +2466,19 @@ void FWFWKWebViewHostApiSetup(id<FlutterBinaryMessenger> binaryMessenger,
         binaryMessenger:binaryMessenger
                   codec:FWFWKWebViewHostApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(setUserAgentForWebViewWithIdentifier:
-                                                                             userAgent:error:)],
+      NSCAssert([api respondsToSelector:@selector
+                     (setCustomUserAgentForWebViewWithIdentifier:userAgent:error:)],
                 @"FWFWKWebViewHostApi api (%@) doesn't respond to "
-                @"@selector(setUserAgentForWebViewWithIdentifier:userAgent:error:)",
+                @"@selector(setCustomUserAgentForWebViewWithIdentifier:userAgent:error:)",
                 api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         NSNumber *arg_identifier = GetNullableObjectAtIndex(args, 0);
         NSString *arg_userAgent = GetNullableObjectAtIndex(args, 1);
         FlutterError *error;
-        [api setUserAgentForWebViewWithIdentifier:arg_identifier
-                                        userAgent:arg_userAgent
-                                            error:&error];
+        [api setCustomUserAgentForWebViewWithIdentifier:arg_identifier
+                                              userAgent:arg_userAgent
+                                                  error:&error];
         callback(wrapResult(nil, error));
       }];
     } else {
@@ -2534,6 +2534,29 @@ void FWFWKWebViewHostApiSetup(id<FlutterBinaryMessenger> binaryMessenger,
                                         inspectable:arg_inspectable
                                               error:&error];
         callback(wrapResult(nil, error));
+      }];
+    } else {
+      [channel setMessageHandler:nil];
+    }
+  }
+  {
+    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
+           initWithName:
+               @"dev.flutter.pigeon.webview_flutter_wkwebview.WKWebViewHostApi.getCustomUserAgent"
+        binaryMessenger:binaryMessenger
+                  codec:FWFWKWebViewHostApiGetCodec()];
+    if (api) {
+      NSCAssert([api respondsToSelector:@selector(customUserAgentForWebViewWithIdentifier:error:)],
+                @"FWFWKWebViewHostApi api (%@) doesn't respond to "
+                @"@selector(customUserAgentForWebViewWithIdentifier:error:)",
+                api);
+      [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
+        NSArray *args = message;
+        NSNumber *arg_identifier = GetNullableObjectAtIndex(args, 0);
+        FlutterError *error;
+        NSString *output = [api customUserAgentForWebViewWithIdentifier:arg_identifier
+                                                                  error:&error];
+        callback(wrapResult(output, error));
       }];
     } else {
       [channel setMessageHandler:nil];