+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/web_kit/web_kit_api_impls.dart
@@ -1109,6 +1109,11 @@ class WKWebViewHostApiImpl extends WKWebViewHostApi {
     );
   }
 
+  /// Calls [getCustomUserAgent] with the ids of the provided object instances.
+  Future<String?> getCustomUserAgentForInstances(WKWebView instance) {
+    return getCustomUserAgent(instanceManager.getIdentifier(instance)!);
+  }
+
   /// Calls [setNavigationDelegate] with the ids of the provided object instances.
   Future<void> setNavigationDelegateForInstances(
     WKWebView instance,