+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,9 +1,17 @@
-## 3.11.0
+## 3.13.0
 
 * Adds support to intercept JavaScript dialog.
   Sees `PlatformWebViewController.setOnJavaScriptAlertDialog`,
   `PlatformWebViewController.setOnJavaScriptConfirmDialog`,
   `PlatformWebViewController.setOnJavaScriptPromptDialog`.
+
+## 3.12.0
+
+* Adds support for `PlatformWebViewController.getUserAgent`.
+
+## 3.11.0
+
+* Adds support to register a callback to receive JavaScript console messages. See `AndroidWebViewController.onConsoleMessage`.
 
 ## 3.10.1
 
@@ -30,7 +38,7 @@
 ## 3.9.2
 
 * Fixes bug where `PlatformWebViewWidget` doesn't rebuild when the controller or PlatformView
-  implementation flag changes. 
+  implementation flag changes.
 
 ## 3.9.1